+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 CHANGELOG
 =========
 
+* service container extension test improvements (by using test case classes
+  from the ``matthiasnoback/symfony-dependency-injection-test`` package)
+
 * do not display misleading pagination information with empty result sets in
   the ``panda:video:list`` command